Move refreshAuthToken to functions
